add form tag in create-post.php file and resolve design issues

+++ admin/dashboard/script/delete-user.php
@@ -5,7 +5,9 @@
     $query1 = "SELECT * FROM users WHERE id = {$DELETE_USER_ID}";
     $result1 = mysqli_query($connection, $query1);
     $row = mysqli_fetch_assoc($result1);
-    unlink("../../uploads/user/".$row['profile_img']);
+    if(!empty($row['profile_img'])){
+        unlink("../../uploads/user/".$row['profile_img']);
+    }
     $query = "DELETE FROM users WHERE id = {$DELETE_USER_ID}";
     $result = mysqli_query($connection, $query);
     if($result){
